+++ projetBlog/show-article.php
@@ -1,5 +1,7 @@
 <?php
-
+require_once __DIR__ . '/database/database.php';
+require_once __DIR__ . '/database/security.php';
+$currentUser = isLoggedIn();
 $articleDB = require_once './database/models/articleDB.php';
 $statement = $pdo->prepare('SELECT * from article WHERE id=:id');
 $_GET = filter_input_array(INPUT_GET, FILTER_SANITIZE_FULL_SPECIAL_CHARS);
@@ -32,10 +34,15 @@ if (!$id) {
                 <h1 class="article-title"><?= $article['title'] ?></h1>
                 <div class="separator"></div>
                 <p class="article-content"><?= $article['content'] ?></p>
-                <div class="action">
-                    <a class="btn btn-secondary" href="/delete-article.php?id=<?= $article['id'] ?>">Supprimer un article</a>
-                    <a class="btn btn-primary" href="form-article.php?id=<?= $article['id'] ?>">Editer l'article</a>
+                <div class="article-author">
+                    <p><?= $article['firstname'] . ' ' . $article['lastname'] ?></p>
                 </div>
+                <?php if ($currentUser && $currentUser['id'] === $article['author']) : ?>
+                    <div class="action">
+                        <a class="btn btn-secondary" href="/delete-article.php?id=<?= $article['id'] ?>">Supprimer un article</a>
+                        <a class="btn btn-primary" href="form-article.php?id=<?= $article['id'] ?>">Editer l'article</a>
+                    </div>
+                <?php endif; ?>
             </div>
         </div>
         <?php require_once('./includes/footer.php') ?>
